+++ partiql-planner/src/main/kotlin/org/partiql/planner/internal/typer/CompilerType.kt
@@ -32,7 +32,7 @@ internal class CompilerType(
         }.toMutableList()
     }
 
-    override fun getLength(): Int {
+    override fun getLength(): Long {
         return _delegate.length
     }
