+++ partiql-spi/src/main/java/org/partiql/spi/types/PTypeWithLongMaxLength.java
@@ -12,8 +12,8 @@ class PTypeWithLongMaxLength extends PType {
     }
 
     @Override
-    public int getLength() {
-        return (int) _maxLength;
+    public long getLength() {
+        return _maxLength;
     }
 
     @Override
