+++ partiql-spi/src/main/kotlin/org/partiql/spi/function/builtins/FnBitLength.kt
@@ -25,7 +25,7 @@ internal val Fn_BIT_LENGTH__CLOB__INT32 = Function.overload(
     name = "bit_length",
     returns = PType.integer(),
     parameters = arrayOf(
-        Parameter("lhs", PType.clob(Int.MAX_VALUE)),
+        Parameter("lhs", PType.clob(Long.MAX_VALUE)),
     ),
 
 ) { args ->
