+++ partiql-spi/src/main/kotlin/org/partiql/spi/function/builtins/FnLikeEscape.kt
@@ -48,9 +48,9 @@ internal val Fn_LIKE_ESCAPE__CLOB_CLOB_CLOB__BOOL = FunctionUtils.hidden(
     name = "like_escape",
     returns = PType.bool(),
     parameters = arrayOf(
-        Parameter("value", PType.clob(Int.MAX_VALUE)),
-        Parameter("pattern", PType.clob(Int.MAX_VALUE)),
-        Parameter("escape", PType.clob(Int.MAX_VALUE)),
+        Parameter("value", PType.clob(Long.MAX_VALUE)),
+        Parameter("pattern", PType.clob(Long.MAX_VALUE)),
+        Parameter("escape", PType.clob(Long.MAX_VALUE)),
     ),
 
 ) { args ->
