+++ partiql-spi/src/main/kotlin/org/partiql/spi/function/builtins/FnTrimTrailing.kt
@@ -24,8 +24,8 @@ internal val Fn_TRIM_TRAILING__STRING__STRING = FunctionUtils.hidden(
 internal val Fn_TRIM_TRAILING__CLOB__CLOB = FunctionUtils.hidden(
 
     name = "trim_trailing",
-    returns = PType.clob(Int.MAX_VALUE),
-    parameters = arrayOf(Parameter("value", PType.clob(Int.MAX_VALUE))),
+    returns = PType.clob(Long.MAX_VALUE),
+    parameters = arrayOf(Parameter("value", PType.clob(Long.MAX_VALUE))),
 
 ) { args ->
     val string = args[0].bytes.toString(Charsets.UTF_8)
