+++ partiql-spi/src/main/kotlin/org/partiql/spi/function/builtins/FnUpper.kt
@@ -23,8 +23,8 @@ internal val Fn_UPPER__STRING__STRING = Function.overload(
 internal val Fn_UPPER__CLOB__CLOB = Function.overload(
 
     name = "upper",
-    returns = PType.clob(Int.MAX_VALUE),
-    parameters = arrayOf(Parameter("value", PType.clob(Int.MAX_VALUE))),
+    returns = PType.clob(Long.MAX_VALUE),
+    parameters = arrayOf(Parameter("value", PType.clob(Long.MAX_VALUE))),
 
 ) { args ->
     val string = args[0].bytes.toString(Charsets.UTF_8)
